Helper functions from similar-advert.js module was moved to new module called 'similar-adverts-utils.js'
